feat: migrate auth to server side refresh/caching
