+++ packages/pieces/community/omnihr/src/lib/actions/get-direct-reports.ts
@@ -1,7 +1,6 @@
 import { createAction, Property } from '@activepieces/pieces-framework';
 import { httpClient, HttpMethod } from '@activepieces/pieces-common';
 import { omnihrAuth } from '../auth';
-import { getAuthHeaders, OmniHrAuth } from '../common/client';
 
 export const getDirectReports = createAction({
   auth: omnihrAuth,
@@ -24,8 +23,11 @@ export const getDirectReports = createAction({
   },
   async run(context) {
     const { system_id } = context.propsValue;
-    const auth = context.auth as OmniHrAuth;
-    const headers = await getAuthHeaders(auth);
+    const headers = {
+      Authorization: `Bearer ${context.auth.access_token}`,
+      'Content-Type': 'application/json',
+      Origin: context.auth.props.origin,
+    };
 
     const directReportsResponse = await httpClient.sendRequest({
       method: HttpMethod.GET,
